added other tests and some rafactoring

+++ src/main/java/com/epam/web/rotten/potatoes/command/impl/admin/DeleteFilmCommand.java
@@ -19,7 +19,10 @@ public class DeleteFilmCommand implements Command {
     @Override
     public CommandResult execute(RequestContext requestContext) throws ServiceException {
         Film film  = (Film) requestContext.getSessionAttribute(FILM);
-        Integer filmId = film.getId();
+        if (film == null) {
+            throw new ServiceException("Incoming parameters are: null");
+        }
+        int filmId = film.getId();
         filmService.remove(filmId);
         return CommandResult.redirect(INDEX_PAGE);
     }
